testing

+++ Model/dataset.py
@@ -32,7 +32,7 @@ class SplitDataset(data.Dataset):
 
     def __init__(self, individuals_file, directory):
         self.directory = directory
-        self.individuals = np.load(os.path.join(directory, individuals_file), dtype="float32")
+        self.individuals = np.load(os.path.join(directory, individuals_file))
         print(self.individuals)
         
     def __len__(self):
